working modularisation of image gen
