+++ index.js
@@ -3,6 +3,9 @@ require('dotenv/config');
 // Import the moderation function from the moderation.js file
 const moderateMessage = require('./moderation');
 
+// Import the generateImage function from the imageGenerator.js file
+const generateImage = require('./imageGenerator');
+
 const axios = require('axios');
 const fs = require('node:fs');
 const path = require('node:path');
@@ -80,41 +83,6 @@ client.on(Events.InteractionCreate, async interaction => {
     }
 
 });
-
-
-//---------------------------------------------------------GENERATE IMAGE FUNCTION---------------------------------------------------------------------
-async function generateImage(prompt) {
-    try {
-        console.log('Generating image with prompt:', prompt);
-
-        const response = await axios.post(
-            'https://api.openai.com/v1/images/generations',
-            {
-                prompt: prompt,
-                n: 1,
-                size: "1024x1024",
-            },
-            {
-                headers: {
-                    'Authorization': `Bearer ${process.env.OPENAI_API_KEY}`,
-                    'Content-Type': 'application/json',
-                },
-            }
-        );
-        console.log('Response received:', response.data);
-
-        if (response.data && response.data.data && response.data.data.length > 0) {
-            const imageData = response.data.data[0].url;
-            return imageData;
-        } else {
-            throw new Error('Failed to generate image');
-        }
-    } catch (error) {
-        console.error('Error generating image:', error);
-        console.error(error);
-        return null;
-    }
-}
 
 // ---------------------------------------------------------RESPONSE GENERATION--------------------------------------------------------------------
 // Listening for events in channels that the bot is in
@@ -197,4 +165,5 @@ client.login(process.env.TOKEN);
 module.exports = {
     client: client,
     handleMessage: handleMessage,
+    generateImage: generateImage,
 };
