+++ test/openai.test.js
@@ -1,5 +1,5 @@
 // Import necessary modules and files
-const { handleMessage } = require('../index.js'); // handleMessage function is exported from index.js
+const { handleMessage, generateImage } = require('../index.js'); // handleMessage function is exported from index.js
 const { test } = require('node:test');
 const assert = require('assert');
 
@@ -53,8 +53,8 @@ test("Dall-E image generation", async () => {
     };
 
     try {
-        // Wait for handleMessage to complete its operation
-        await handleMessage(testMsg);
+        // Wait for generateImage to complete its operation
+        await generateImage(testMsg);
         // If handleMessage completes without errors, and send function logic passes, test is considered successful
         console.log("Test passed: Image generation and response successful.");
     } catch (error) {
